relocate oyster tests; break them up
